Refactor login and logout pages
I also disable WalletConnect – does not work anyway.
