+++ app/package.json
@@ -122,9 +122,6 @@
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.3",
     "@tryghost/content-api": "^1.4.13",
-    "@walletconnect/client": "^1.3.4",
-    "@walletconnect/qrcode-modal": "^1.3.4",
-    "@walletconnect/utils": "^1.3.4",
     "@wojtekmaj/enzyme-adapter-react-17": "^0.6.6",
     "abi-decoder": "^2.4.0",
     "axios": "^0.21.2",
