--- app/src/auth/components/AuthLayout/authLayout.pcss
@@ -1,5 +1,0 @@
-.html,
-.html body {
-  background: #F8F8F8;
-  color: #323232;
-}
