+++ app/src/pages/LoginPage/adapters/WalletConnect.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import { SignInMethod } from '@streamr/streamr-layout'
+
+const WalletConnect = {
+    id: 'walletConnect',
+    label: 'WalletConnect',
+    icon: <SignInMethod.Icon.WalletConnect />,
+    async connect() {
+        throw new Error('Not implemented')
+    },
+}
+
+export default WalletConnect
